add versioning to memmap.json files
The esp-idf-size tool will offer a basic API, allowing third-party apps
to use the JSON memory map. To facilitate this, a version has been added
to the memory map. To prevent the need for generating new reference
memory maps, the version is manually added to existing ones. The format
remains unchanged except for the addition of the version.

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>